fix: 서버의 Price 타입 변경 (Int->Double)에 따른 수정 #3
- Product 타입의 price, bargainPrice, discountedPrice 프로퍼티 타입 변경 (Int -> Double)
- Double 타입 익스텐션 추가

+++ OpenMarket/OpenMarket/View/ListProductCell.swift
@@ -140,9 +140,9 @@ class ListProductCell: UICollectionViewCell, ProductCellProtocol {
         changeStockLabel(by: data.stock)
     }
     
-    private func changePriceAndDiscountedPriceLabel(price: Int,
-                                                    discountedPrice: Int,
-                                                    bargainPrice: Int,
+    private func changePriceAndDiscountedPriceLabel(price: Double,
+                                                    discountedPrice: Double,
+                                                    bargainPrice: Double,
                                                     currency: Currency) {
         if discountedPrice == 0 {
             priceLabel.attributedText = nil
